requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.99.1
fastdeploy-python==1.0.7
fastdeploy-tools==0.0.5
h11==0.14.0
idna==3.4
numpy==1.25.0
opencv-python==4.8.0.74
pydantic==1.10.11
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pywin32==306
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing-extensions==4.7.1
urllib3==2.0.3
uvicorn==0.16.0
